Split addons in parser and addon
