+++ matrixctl/addons/deluser/addon.py
@@ -21,10 +21,13 @@ from __future__ import annotations
 
 import logging
 
-from argparse import ArgumentParser
-from argparse import _SubParsersAction as SubParsersAction
+from argparse import Namespace
 
-from argparse_addon_manager.addon_manager import AddonManager
+
+from matrixctl.errors import InternalResponseError
+from matrixctl.handlers.api import RequestBuilder
+from matrixctl.handlers.api import request
+from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
@@ -34,26 +37,36 @@ __email__: str = "Michael@MichaelSasser.org"
 logger = logging.getLogger(__name__)
 
 
-@AddonManager.add_subparser
-def subparser_deluser(subparsers: SubParsersAction) -> None:
-    """Create a subparser for the ``matrixctl deluser`` command.
+def addon(arg: Namespace, yaml: YAML) -> int:
+    """Delete a user from the the matrix instance.
 
     Parameters
     ----------
-    subparsers : argparse._SubParsersAction
-        The object which is returned by
-        ``parser.add_subparsers()``.
+    arg : argparse.Namespace
+        The ``Namespace`` object of argparse's ``parse_args()``
+    yaml : matrixctl.handlers.yaml.YAML
+        The configuration file handler.
 
     Returns
     -------
-    None
+    err_code : int
+        Non-zero value indicates error code, or zero on success.
 
     """
-    parser: ArgumentParser = subparsers.add_parser(
-        "deluser", help="Deletes a user"
+    req: RequestBuilder = RequestBuilder(
+        token=yaml.get("api", "token"),
+        domain=yaml.get("api", "domain"),
+        path=f"deactivate/@{arg.user}:{yaml.get('api','domain')}",
+        api_version="v1",
+        method="POST",
+        data={"erase": True},
     )
-    parser.add_argument("user", help="The username to delete")
-    parser.set_defaults(addon="deluser")
+    try:
+        request(req)
+    except InternalResponseError:
+        logger.error("The user was not deleted.")
+
+    return 0
 
 
 # vim: set ft=python :
