+++ matrixctl/addons/deploy/parser.py
@@ -22,13 +22,9 @@ from __future__ import annotations
 import logging
 
 from argparse import ArgumentParser
-from argparse import Namespace
 from argparse import _SubParsersAction as SubParsersAction
 
 from argparse_addon_manager.addon_manager import AddonManager
-
-from matrixctl.handlers.ansible import ansible_run
-from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
@@ -63,31 +59,7 @@ def subparser_deploy(subparsers: SubParsersAction) -> None:
         action="store_true",
         help="Start/Restart after the deployment",
     )
-    parser.set_defaults(func=deploy)
-
-
-def deploy(arg: Namespace, yaml: YAML) -> int:
-    """Deploy the ansible playbook.
-
-    Parameters
-    ----------
-    arg : argparse.Namespace
-        The ``Namespace`` object of argparse's ``parse_args()``
-    yaml : matrixctl.handlers.yaml.YAML
-        The configuration file handler.
-
-    Returns
-    -------
-    err_code : int
-        Non-zero value indicates error code, or zero on success.
-
-    """
-    ansible_run(
-        playbook=yaml.get("ansible", "playbook"),
-        tags="setup-all,start" if arg.start else "setup-all",
-    )
-
-    return 0
+    parser.set_defaults(addon="deploy")
 
 
 # vim: set ft=python :
