+++ matrixctl/addons/stop/parser.py
@@ -20,13 +20,9 @@
 from __future__ import annotations
 
 from argparse import ArgumentParser
-from argparse import Namespace
 from argparse import _SubParsersAction as SubParsersAction
 
 from argparse_addon_manager.addon_manager import AddonManager
-
-from matrixctl.handlers.ansible import ansible_run
-from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
@@ -50,27 +46,7 @@ def subparser_stop(subparsers: SubParsersAction) -> None:
     parser: ArgumentParser = subparsers.add_parser(
         "stop", help="Stops all OCI containers"
     )
-    parser.set_defaults(func=stop)
-
-
-def stop(_: Namespace, yaml: YAML) -> int:
-    """Stop the OCI containers.
-
-    Parameters
-    ----------
-    arg : argparse.Namespace
-        The ``Namespace`` object of argparse's ``parse_args()``.
-    yaml : matrixctl.handlers.yaml.YAML
-        The configuration file handler.
-
-    Returns
-    -------
-    err_code : int
-        Non-zero value indicates error code, or zero on success.
-
-    """
-    ansible_run(yaml.get("ansible", "playbook"), tags="stop")
-    return 0
+    parser.set_defaults(addon="stop")
 
 
 # vim: set ft=python :
