+++ matrixctl/addons/update/addon.py
@@ -15,43 +15,40 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-"""Use this module to add the ``version`` subcommand to ``matrixctl``."""
+"""Use this module to add the ``update`` subcommand to ``matrixctl``."""
 
 from __future__ import annotations
 
-import logging
+from argparse import Namespace
 
-from argparse import ArgumentParser
-from argparse import _SubParsersAction as SubParsersAction
-
-from argparse_addon_manager.addon_manager import AddonManager
+from matrixctl.handlers.vcs import VCS
+from matrixctl.handlers.yaml import YAML
 
 
 __author__: str = "Michael Sasser"
 __email__: str = "Michael@MichaelSasser.org"
 
 
-logger = logging.getLogger(__name__)
-
-
-@AddonManager.add_subparser
-def subparser_version(subparsers: SubParsersAction) -> None:
-    """Create a subparser for the ``matrixctl version`` command.
+def addon(_: Namespace, yaml: YAML) -> int:
+    """Update the synapse playbook with git.
 
     Parameters
     ----------
-    subparsers : argparse._SubParsersAction
-        The object which is returned by ``parser.add_subparsers()``.
+    arg : argparse.Namespace
+        The ``Namespace`` object of argparse's ``parse_args()``.
+    yaml : matrixctl.handlers.yaml.YAML
+        The configuration file handler.
 
     Returns
     -------
-    None
+    err_code : int
+        Non-zero value indicates error code, or zero on success.
 
     """
-    parser: ArgumentParser = subparsers.add_parser(
-        "version", help="Get the version of the Synapse instance"
-    )
-    parser.set_defaults(addon="version")
+    git: VCS = VCS(yaml.get("synapse", "playbook"))
+    git.pull()
+
+    return 0
 
 
 # vim: set ft=python :
