+++ matrixctl/addons/user/addon.py
@@ -1,0 +1,245 @@
+#!/usr/bin/env python
+# matrixctl
+# Copyright (c) 2020  Michael Sasser <Michael@MichaelSasser.org>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Use this module to add the ``user`` subcommand to ``matrixctl``."""
+
+from __future__ import annotations
+
+import datetime
+import logging
+import sys
+
+from argparse import Namespace
+from typing import Any
+
+from tabulate import tabulate
+
+from matrixctl.errors import InternalResponseError
+from matrixctl.handlers.api import RequestBuilder
+from matrixctl.handlers.api import request
+from matrixctl.handlers.yaml import YAML
+from matrixctl.print_helpers import human_readable_bool
+from matrixctl.typehints import JsonDict
+
+
+__author__: str = "Michael Sasser"
+__email__: str = "Michael@MichaelSasser.org"
+
+
+logger = logging.getLogger(__name__)
+
+
+def make_human_readable(
+    k: str, user_dict: dict[str, str], len_domain: int
+) -> tuple[str, str]:
+    """Make a key/value pair of a ``user`` (line) human readable, by modifying.
+
+    Notes
+    -----
+    This function is used as helper by ``matrixctl.user.generate_user_tables``.
+
+    Parameters
+    ----------
+    k : str
+        The key
+    user_dict : `dict` [`str`, `Any`]
+        The line as dict, a JSON string which was converted to a Python
+        dictionary. (This is not a ``Collections.UserDict``)
+    len_domain : int
+        The length in characters of the domain.
+
+    Returns
+    -------
+    err_code : int
+        Non-zero value indicates error code, or zero on success.
+
+    """
+
+    key: str | None = None
+    value: str
+
+    if k == "name":
+        value = str(user_dict[k][1:-len_domain])
+    elif k == "is_guest":
+        value = human_readable_bool(user_dict[k])
+        key = "Guest"
+    elif k in ("admin", "deactivated"):
+        value = human_readable_bool(user_dict[k])
+    elif k.endswith("_ts"):
+        value = str(
+            datetime.datetime.fromtimestamp(float(user_dict[k]))
+        )  # UTC?
+    elif k.endswith("_at"):
+        value = str(
+            datetime.datetime.fromtimestamp(float(user_dict[k]) / 1000.0)
+        )  # UTC?
+
+    else:
+        value = user_dict[k]
+
+    if key is None:
+        key = k.replace("_", " ").title()
+
+    return key, value
+
+
+# TODO: JSON Type?
+# TODO: Returns type wrong
+def generate_user_tables(
+    user_dict: dict[str, Any], len_domain: int
+) -> list[list[tuple[str, str]]]:
+    """Generate a main user table and threepid user tables.
+
+    The function gnerates first a main user table and then for every threepid
+    a additional table from a ``user_dict``.
+    It renames and makes the output human readable.
+
+    Notes
+    -----
+    This function is a recursive function.
+
+    Parameters
+    ----------
+    user_dict : `dict` [`str`, `Any`]
+        The line as dict, a JSON string which was converted to a Python
+        dictionary. (This is not a ``Collections.UserDict``)
+    len_domain : int
+        The length in characters of the domain.
+
+    Returns
+    -------
+    err_code : int
+        A list in the format: ``[[main], threepids_0, ... ,threepids_n]``
+
+    """
+
+    table: list[list[tuple[str, str]]] = [[]]
+
+    for k in user_dict:
+        if k == "errcode":
+            logger.error("There is no user with that username.")
+            sys.exit(1)
+
+        if k == "threepids":
+            for tk in user_dict[k]:
+                ret: list[list[tuple[str, str]]] = generate_user_tables(
+                    tk, len_domain
+                )
+                table.append(ret[0])
+
+            continue  # Don't add threepids to "table"
+
+        table[0].append(make_human_readable(k, user_dict, len_domain))
+
+    return table
+
+
+def addon(arg: Namespace, yaml: YAML) -> int:
+    """List information about an registered user.
+
+    It uses the admin API to get a python dictionary with the information.
+    The ``generate_user_tables`` function makes the information human readable.
+    The Python package ``tabulate`` renders the table as shown below, if
+    everything works well.
+
+    Examples
+    --------
+    .. code-block:: console
+
+       $ matrixctl user dwight
+       User:
+       +----------------------------+--------------------------------------+
+       | Name                       | dwight                               |
+       | Password Hash              | $2b$12$9DUNderm1ffL1NincPap3RC       |
+       |                            | ompaNY1725.slOUghAvEnu5cranT0n       |
+       | Guest                      | False                                |
+       | Admin                      | True                                 |
+       | Consent Version            |                                      |
+       | Consent Server Notice Sent |                                      |
+       | Appservice Id              |                                      |
+       | Creation Ts                | 2020-04-14 13:04:21                  |
+       | User Type                  |                                      |
+       | Deactivated                | False                                |
+       | Displayname                | Dwight Schrute                       |
+       | Avatar Url                 | mxc://dunder-mifflin.com/sCr4        |
+       |                            | nt0nsr4ng13rW45Cr33d                 |
+       +----------------------------+--------------------------------------+
+
+       Threepid:
+       +--------------+-----------------------------------+
+       | Medium       | email                             |
+       | Address      | dwight_schrute@dunder-mifflin.com |
+       | Validated At | 2020-04-14 15:30:21.123000        |
+       | Added At     | 2020-04-14 15:29:19.100000        |
+       +--------------+-----------------------------------+
+
+    If the user does not exist, the return looks like:
+
+    .. code-block:: console
+
+       $ matrixctl user mose
+       2020-04-14 13:58:13 - ERROR - The request was not successful.
+       2020-04-14 13:58:13 - ERROR - There is no user with that username.
+
+    Parameters
+    ----------
+    arg : argparse.Namespace
+        The ``Namespace`` object of argparse's ``parse_args()``.
+    yaml : matrixctl.handlers.yaml.YAML
+        The configuration file handler.
+
+    Returns
+    -------
+    err_code : int
+        Non-zero value indicates error code, or zero on success.
+
+    """
+    req: RequestBuilder = RequestBuilder(
+        token=yaml.get("api", "token"),
+        domain=yaml.get("api", "domain"),
+        path=f'users/@{arg.user}:{yaml.get("api","domain")}',
+    )
+
+    try:
+        user_dict: JsonDict = request(req).json()
+    except InternalResponseError:
+        logger.critical("Could not receive the user information")
+
+        return 1
+
+    len_domain = len(yaml.get("api", "domain")) + 1  # 1 for :
+    user_tables = generate_user_tables(user_dict, len_domain)
+
+    logger.debug(f"User: {user_tables=}")
+
+    for num, table in enumerate(user_tables):
+
+        if num < 1:
+            print("User:")
+        else:
+            print("\nThreepid:")
+        print(
+            tabulate(
+                table,
+                tablefmt="psql",
+            )
+        )
+
+    return 0
+
+
+# vim: set ft=python :
