+++ matrixctl/addons/version/addon.py
@@ -1,0 +1,79 @@
+#!/usr/bin/env python
+# matrixctl
+# Copyright (c) 2020  Michael Sasser <Michael@MichaelSasser.org>
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""Use this module to add the ``version`` subcommand to ``matrixctl``."""
+
+from __future__ import annotations
+
+import logging
+
+from argparse import Namespace
+
+from matrixctl.errors import InternalResponseError
+from matrixctl.handlers.api import RequestBuilder
+from matrixctl.handlers.api import request
+from matrixctl.handlers.yaml import YAML
+from matrixctl.typehints import JsonDict
+
+
+__author__: str = "Michael Sasser"
+__email__: str = "Michael@MichaelSasser.org"
+
+
+logger = logging.getLogger(__name__)
+
+
+def addon(_: Namespace, yaml: YAML) -> int:
+    """Get the version of the Synapse instance.
+
+    Parameters
+    ----------
+    arg : argparse.Namespace
+        The ``Namespace`` object of argparse's ``parse_args()``.
+        (Unused in this function)
+    yaml : matrixctl.handlers.yaml.YAML
+        The configuration file handler.
+
+    Returns
+    -------
+    err_code : int
+        Non-zero value indicates error code, or zero on success.
+
+    """
+    req: RequestBuilder = RequestBuilder(
+        token=yaml.get("api", "token"),
+        domain=yaml.get("api", "domain"),
+        path="server_version",
+        api_version="v1",
+    )
+    try:
+        response: JsonDict = request(req).json()
+    except InternalResponseError:
+        logger.critical("Could not get the server sersion.")
+
+        return 1
+    logger.debug(f"{response=}")
+    try:
+        print(f"Server Version: {response['server_version']}")
+        print(f"Python Version: {response['python_version']}")
+    except KeyError:
+        logger.error("MatrixCtl was not able to read the server version.")
+
+    return 0
+
+
+# vim: set ft=python :
